+++ config/karma.base.js
@@ -16,7 +16,6 @@
 
 const karma = require('karma');
 const path = require('path');
-const mochaConfig = require('./mocha.base');
 const webpackTestConfig = require('./webpack.test');
 const { argv } = require('yargs');
 
@@ -102,7 +101,9 @@ const config = {
   singleRun: false,
 
   client: {
-    mocha: mochaConfig,
+    mocha: {
+      opts: `${__dirname}/mocha.browser.opts`
+    },
 
     // Pass through --grep option to filter the tests that run.
     args: argv.grep ? ['--grep', argv.grep] : []