+++ packages/app/rollup.config.js
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-import resolve from 'rollup-plugin-node-resolve';
 import typescript from 'rollup-plugin-typescript2';
-import commonjs from 'rollup-plugin-commonjs';
 import replace from 'rollup-plugin-replace';
 import pkg from './package.json';
 
@@ -31,14 +29,13 @@ const plugins = [
     values: {
       JSCORE_VERSION: firebasePkg.version
     }
-  }),
-  resolve(),
-  commonjs()
+  })
 ];
 
-const external = Object.keys(
+const deps = Object.keys(
   Object.assign({}, pkg.peerDependencies, pkg.dependencies)
 );
+
 export default [
   {
     input: 'index.ts',
@@ -47,7 +44,7 @@ export default [
       { file: pkg.module, format: 'es' }
     ],
     plugins,
-    external
+    external: id => deps.some(dep => id === dep || id.startsWith(`${dep}/`))
   },
   {
     input: 'index.node.ts',
@@ -56,7 +53,7 @@ export default [
       format: 'cjs'
     },
     plugins,
-    external
+    external: id => deps.some(dep => id === dep || id.startsWith(`${dep}/`))
   },
   {
     input: 'index.rn.ts',
@@ -65,6 +62,9 @@ export default [
       format: 'cjs'
     },
     plugins,
-    external: [...external, 'react-native']
+    external: id =>
+      [...deps, 'react-native'].some(
+        dep => id === dep || id.startsWith(`${dep}/`)
+      )
   }
 ];