+++ packages/firestore/karma.conf.js
@@ -60,14 +60,12 @@ function getFirestoreSettings(argv) {
   if (argv.local) {
     return {
       host: 'localhost:8080',
-      ssl: false,
-      timestampsInSnapshots: true
+      ssl: false
     };
   } else {
     return {
       host: 'firestore.googleapis.com',
-      ssl: true,
-      timestampsInSnapshots: true
+      ssl: true
     };
   }
 }