+++ packages/firestore/src/local/indexeddb_mutation_queue.ts
@@ -25,6 +25,7 @@ import { BATCHID_UNKNOWN, MutationBatch } from '../model/mutation_batch';
 import { ResourcePath } from '../model/path';
 import { assert, fail } from '../util/assert';
 import { primitiveComparator } from '../util/misc';
+import { SortedMap } from '../util/sorted_map';
 import { SortedSet } from '../util/sorted_set';
 
 import * as EncodedResourcePath from './encoded_resource_path';
@@ -46,6 +47,8 @@ import { MutationQueue } from './mutation_queue';
 import { PersistenceTransaction, ReferenceDelegate } from './persistence';
 import { PersistencePromise } from './persistence_promise';
 import { SimpleDbStore, SimpleDbTransaction } from './simple_db';
+
+import { AnyJs } from '../../src/util/misc';
 
 /** A mutation queue for a specific user, backed by IndexedDB. */
 export class IndexedDbMutationQueue implements MutationQueue {
@@ -122,13 +125,6 @@ export class IndexedDbMutationQueue implements MutationQueue {
     streamToken: ProtoByteString
   ): PersistencePromise<void> {
     return this.getMutationQueueMetadata(transaction).next(metadata => {
-      const batchId = batch.batchId;
-      assert(
-        batchId > metadata.lastAcknowledgedBatchId,
-        'Mutation batchIDs must be acknowledged in order'
-      );
-
-      metadata.lastAcknowledgedBatchId = batchId;
       metadata.lastStreamToken = convertStreamToken(streamToken);
 
       return mutationQueuesStore(transaction).put(metadata);
@@ -244,11 +240,7 @@ export class IndexedDbMutationQueue implements MutationQueue {
     batchId: BatchId
   ): PersistencePromise<MutationBatch | null> {
     return this.getMutationQueueMetadata(transaction).next(metadata => {
-      // All batches with batchId <= this.metadata.lastAcknowledgedBatchId have
-      // been acknowledged so the first unacknowledged batch after batchID will
-      // have a batchID larger than both of these values.
-      const nextBatchId =
-        Math.max(batchId, metadata.lastAcknowledgedBatchId) + 1;
+      const nextBatchId = batchId + 1;
 
       const range = IDBKeyRange.lowerBound([this.userId, nextBatchId]);
       let foundBatch: MutationBatch | null = null;
@@ -339,7 +331,7 @@ export class IndexedDbMutationQueue implements MutationQueue {
 
   getAllMutationBatchesAffectingDocumentKeys(
     transaction: PersistenceTransaction,
-    documentKeys: DocumentKeySet
+    documentKeys: SortedMap<DocumentKey, AnyJs>
   ): PersistencePromise<MutationBatch[]> {
     let uniqueBatchIDs = new SortedSet<BatchId>(primitiveComparator);