+++ packages/firestore/test/integration/browser/webchannel.test.ts
@@ -42,7 +42,7 @@ describeFn('WebChannel', () => {
     it('includes project ID and database ID', () => {
       const url = makeUrl('Commit', {});
       expect(url).to.equal(
-        'http://example.com/v1beta1/projects/testproject/' +
+        'http://example.com/v1/projects/testproject/' +
           'databases/(default)/documents:commit'
       );
     });