+++ packages/firestore/test/unit/local/test_mutation_queue.ts
@@ -23,6 +23,7 @@ import { DocumentKeySet } from '../../../src/model/collections';
 import { DocumentKey } from '../../../src/model/document_key';
 import { Mutation } from '../../../src/model/mutation';
 import { MutationBatch } from '../../../src/model/mutation_batch';
+import { SortedMap } from '../../../src/util/sorted_map';
 
 /**
  * A wrapper around a MutationQueue that automatically creates a
@@ -144,13 +145,18 @@ export class TestMutationQueue {
   getAllMutationBatchesAffectingDocumentKeys(
     documentKeys: DocumentKeySet
   ): Promise<MutationBatch[]> {
+    let keyMap = new SortedMap<DocumentKey, null>(DocumentKey.comparator);
+    documentKeys.forEach(key => {
+      keyMap = keyMap.insert(key, null);
+    });
+
     return this.persistence.runTransaction(
       'getAllMutationBatchesAffectingDocumentKeys',
       'readonly',
       txn => {
         return this.queue.getAllMutationBatchesAffectingDocumentKeys(
           txn,
-          documentKeys
+          keyMap
         );
       }
     );